Reddit Pict to FB and IG
This file will scrape reddit for hot post with pictures then will post that into FB and IG.

Creates tags from the picture tittle by splitting each word and adding # at the beginning of each word including a small divider in the middle in order to hide the tags from picture tittle.

It is also using information from a config file for FB and IG tokens and IDs, PRAW is unable to get that info, it is possible to edit the praw.ini but for the code we will just enter the reddit app info in the code.