Reddit Video to FB and IG
Included posting to IG, the problem is the format, since we are uploading from reddit,tube sometimes IG will not accept the format of the video and it will cause an error.

We are not manipulating the video locally we are just passing a direct download link from reddit tube for the video posting, so far it posts 2/5 from the subreddit list I am using.

Moved creds and Subreddit List to Config File for easier maintenance.